@lasithdilshan20 - "Update package versions and refactor storeValue command
This commit updates the 'cypress-plugin-store' version to 1.3.3 and the Cypress version to 12.16.0 for compatibility improvements. It also refactors the 'storeValue' command in 'spec.cy.js' and 'store-value.js' to simplify and improve readability. Additionally, the project ID in 'cypress.config.js' was updated for better project tracking and the 'storeValue' function was optimized with a different approach for accessing 'text' and 'val' methods to adjust the value extraction process for both CSS and Xpath selectors. "