Update cypress version 13.4.x versions
This commit updates the versions of several dependencies in package-lock.json. The dependencies were updated to get the latest fixes and features. Updated dependencies include cypress, @cypress/request, @types/node, call-bind, function-bind. The objective is to keep the project updated with the latest stable version of each dependency, ensuring the best possible performance and security.